+++ apps/server/src/modules/database/index.ts
@@ -1,1 +1,2 @@
 export * from './mongo-memory-database';
+export * from './setup-entities';