+++ CHANGELOG.md
@@ -12,8 +12,16 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 ### Added
 
 - BC-394 - send email wishes from federal states also to ticketsystem
+- BC-391 - display special characters properly
+
+### Fixed
+
+- BC-436 - Fix univention LDAP strategy
 
 ### Changed
+
+- BC-487 - update mongodb-memory-server from 6.9.6 to 7.4.4
+- BC-490 - setup entities for tests without needing a database
 
 ## [26.12.0] - 2021-10-12
 
@@ -22,6 +30,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 - BC-388 - add simple script for end sjw
 - BC-319 - add route to change the position of a course dashboard element
 - BC-343 - run unstable e2e tests on label `run unstable tests`
+- BC-6   - Show prepared tasks (drafts) for teachers
 
 ### Changed
 
@@ -30,6 +39,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 - BC-237 - Clean up global entity registration
 - BC-331 - BC-354 - rename ansible variables for OnePassword
 - BC-269 - check in firstLogin that students can only pass with existing birthdate
+- BC-154 - remove lesson entity from task repository
 - BC-231 - upgraded nestjs@7 to nestjs@8
 
 ### Fixed